툴 내구도 구현 중

+++ Assets/02_Script/Item/ScriptableObject/EquippableItemData.cs
@@ -14,8 +14,8 @@ public class EquippableItemData : ItemData
     [Header("EquipmentItemData")]
     [SerializeField] EquipmentSlot equipSlot;
 
-    [SerializeField] public float maxDurability;
-    [SerializeField] public float currentDurability;
+    [SerializeField] public int maxDurability;
+    [SerializeField] public int currentDurability;
 
     public EquipmentSlot EquipSlot { get { return equipSlot; } }
 }
